requirements.txt
azure-functions
requests
openpyxl
sqlalchemy
pyodbc